Connectivity: start common.cc and connectit.cc files
